Move useStation into @neaps/tide-pridictor
