+++ .changeset/free-pumas-occur.md
@@ -1,0 +1,6 @@
+---
+"@neaps/tide-predictor": minor
+"neaps": minor
+---
+
+Moved `useStation` into @neaps/tide-predictor so it can be used without the heavy dependency of @neaps/tide-database.
