+++ packages/tide-predictor/test/index.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import mockConstituents from "./_mocks/constituents.js";
-import tidePrediction from "../src/index.js";
+import { createTidePredictor } from "../src/index.js";
 
 const startDate = new Date("2019-09-01T00:00:00Z");
 const endDate = new Date("2019-09-01T06:00:00Z");
@@ -10,14 +10,14 @@ describe("Tidal station", () => {
     let stationCreated = true;
 
     try {
-      tidePrediction(mockConstituents);
+      createTidePredictor(mockConstituents);
     } catch {
       stationCreated = false;
     }
     expect(stationCreated).toBe(true);
 
     try {
-      tidePrediction(mockConstituents);
+      createTidePredictor(mockConstituents);
     } catch {
       stationCreated = false;
     }
@@ -25,7 +25,7 @@ describe("Tidal station", () => {
   });
 
   it("it predicts the tides in a timeline", () => {
-    const results = tidePrediction(mockConstituents).getTimelinePrediction({
+    const results = createTidePredictor(mockConstituents).getTimelinePrediction({
       start: startDate,
       end: endDate,
     });
@@ -36,7 +36,7 @@ describe("Tidal station", () => {
   });
 
   it("it predicts the tides in a timeline with time fidelity", () => {
-    const results = tidePrediction(mockConstituents).getTimelinePrediction({
+    const results = createTidePredictor(mockConstituents).getTimelinePrediction({
       start: startDate,
       end: endDate,
       timeFidelity: 60,
@@ -48,7 +48,7 @@ describe("Tidal station", () => {
   });
 
   it("it predicts the tidal extremes", () => {
-    const results = tidePrediction(mockConstituents).getExtremesPrediction({
+    const results = createTidePredictor(mockConstituents).getExtremesPrediction({
       start: startDate,
       end: endDate,
     });
@@ -56,7 +56,7 @@ describe("Tidal station", () => {
   });
 
   it("it predicts the tidal extremes with high fidelity", () => {
-    const results = tidePrediction(mockConstituents).getExtremesPrediction({
+    const results = createTidePredictor(mockConstituents).getExtremesPrediction({
       start: startDate,
       end: endDate,
       timeFidelity: 60,
@@ -65,17 +65,43 @@ describe("Tidal station", () => {
   });
 
   it("it fetches a single water level", () => {
-    const result = tidePrediction(mockConstituents).getWaterLevelAtTime({
+    const result = createTidePredictor(mockConstituents).getWaterLevelAtTime({
       time: startDate,
     });
     expect(result.level).toBeCloseTo(-1.46903456, 4);
   });
 
   it("it adds offset phases", () => {
-    const results = tidePrediction(mockConstituents, {
+    const results = createTidePredictor(mockConstituents, {
       offset: 3,
     }).getExtremesPrediction({ start: startDate, end: endDate });
 
     expect(results[0].level).toBeCloseTo(1.32716067, 4);
+  });
+
+  it("equivalent instants in different timezones yield identical extremes", () => {
+    const predictor = createTidePredictor(mockConstituents);
+
+    const utc = { start: new Date("2019-09-01T00:00:00Z"), end: new Date("2019-09-01T06:00:00Z") };
+    const newYork = {
+      start: new Date("2019-08-31T20:00:00-04:00"),
+      end: new Date("2019-09-01T02:00:00-04:00"),
+    };
+    const tokyo = {
+      start: new Date("2019-09-01T09:00:00+09:00"),
+      end: new Date("2019-09-01T15:00:00+09:00"),
+    };
+
+    const baseline = predictor.getExtremesPrediction(utc);
+    const ny = predictor.getExtremesPrediction(newYork);
+    const jp = predictor.getExtremesPrediction(tokyo);
+
+    [ny, jp].forEach((result) => {
+      expect(result.length).toBe(baseline.length);
+      result.forEach((extreme, index) => {
+        expect(extreme.time.valueOf()).toBe(baseline[index].time.valueOf());
+        expect(extreme.level).toBeCloseTo(baseline[index].level, 6);
+      });
+    });
   });
 });
